refactor folder structure by moving out of src folder
