+++ Source/DesignSystem/.storybook/main.js
@@ -2,8 +2,6 @@ const path = require('path');
 
 module.exports = {
   "stories": [
-    "../src/**/*.stories.mdx",
-    "../src/**/*.stories.@(js|jsx|ts|tsx)",
     "../atoms/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
