+++ Source/DesignSystem/tsconfig.json
@@ -22,5 +22,5 @@
   },
   "include": [
     "src"
-, "atoms/Button", "Button"  ]
+, "atoms/Button", "Button", "App.tsx"  ]
 }
